+++ src/dev-server/util.ts
@@ -85,8 +85,13 @@ export function isSimpleText(filePath: string) {
 }
 
 
-export function isStaticDevClient(req: d.HttpRequest) {
-  return req.pathname.startsWith(DEV_SERVER_URL);
+export function isDevClient(pathname: string) {
+  return pathname.startsWith(DEV_SERVER_URL);
+}
+
+
+export function isOpenInEditor(pathname: string) {
+  return pathname === OPEN_IN_EDITOR_URL;
 }
 
 
@@ -103,6 +108,8 @@ export function isDevServerClient(pathname: string) {
 export const DEV_SERVER_URL = '/~dev-server';
 
 export const UNREGISTER_SW_URL = `${DEV_SERVER_URL}-init`;
+
+export const OPEN_IN_EDITOR_URL = `${DEV_SERVER_URL}-open-in-editor`;
 
 
 export function shouldCompress(devServerConfig: d.DevServerConfig, req: d.HttpRequest, contentLength: number) {